% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-roi-model.R
\name{as_ROI_model}
\alias{as_ROI_model}
\title{Export to ROI::OP}
\usage{
as_ROI_model(model)
}
\arguments{
\item{model}{an ompr model}
}
\value{
an object of S3 class `ROI::OP`
}
\description{
This function can be used to transform an ompr model to a ROI::OP object.
}
