% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ompr.roi-package.R
\docType{package}
\name{ompr.roi-package}
\alias{ompr.roi}
\alias{ompr.roi-package}
\title{A Solver for 'ompr' that Uses the R Optimization Infrastructure ('ROI')}
\description{
A solver for 'ompr' based on the R Optimization Infrastructure ('ROI').
The package makes all solvers in 'ROI' available to solve 'ompr' models. Please see the
'ompr' website <https://dirkschumacher.github.io/ompr> and package docs for examples on
how to use it.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/dirkschumacher/ompr.roi}
  \item Report bugs at \url{https://github.com/dirkschumacher/ompr.roi/issues}
}

}
\author{
\strong{Maintainer}: Dirk Schumacher \email{mail@dirk-schumacher.net}

}
