% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ompr.roi.R
\docType{package}
\name{ompr.roi}
\alias{ompr.roi}
\alias{ompr.roi-package}
\title{A Solver for 'ompr' that Uses the R Optimization Infrastructure ('ROI')}
\description{
A solver for 'ompr' based on the R Optimization Infrastructure ('ROI').
The package makes all solvers in 'ROI' available to solve 'ompr' models. Please see the
'ompr' website <https://dirkschumacher.github.io/ompr> and package docs for examples on
how to use it.
}
