% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/with-roi.R
\name{with_ROI}
\alias{with_ROI}
\title{Configures a solver based on ROI}
\usage{
with_ROI(solver, ...)
}
\arguments{
\item{solver}{the ROI solver name}

\item{...}{optional parameters passed to ROI_solve}
}
\value{
a function: Model -> Solution
}
\description{
Configures a solver based on ROI
}
\examples{
library(magrittr)
library(ompr)
library(ROI)
library(ROI.plugin.glpk)
add_variable(MIPModel(), x, type = "continuous") \%>\%
 set_objective(x, direction = "max") \%>\%
 add_constraint(x >= 0) \%>\%
 solve_model(with_ROI(solver = "glpk"))

}

