% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/with-roi.R
\name{with_ROI}
\alias{with_ROI}
\title{Configures a solver based on ROI}
\usage{
with_ROI(solver, verbose = FALSE)
}
\arguments{
\item{solver}{the ROI solver name}

\item{verbose}{should status information be printed}
}
\value{
a function: Model -> Solution
}
\description{
Configures a solver based on ROI
}

