% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/with-roi.R
\name{with_ROI}
\alias{with_ROI}
\title{Configures a solver based on 'ROI'}
\usage{
with_ROI(solver, ...)
}
\arguments{
\item{solver}{the 'ROI' solver name (character vector of length 1)}

\item{...}{optional parameters passed to ROI_solve

ROI only returns two status codes, optimal or infeasible. In particular,
this means that when the search is terminated by the user (e.g. due to a time limit),
ROI flags that solution as 'infeasible'.
It is currently not possible for `ompr.roi` to decide, if a solution
is feasible, but not optimal. Therefore only optimal solutions are returned.}
}
\value{
a function: Model -> Solution that can be used
together with \code{\link[ompr]{solve_model}}.
}
\description{
This function makes all solvers in the R package 'ROI' available to solve
'ompr' models.
}
\examples{
library(magrittr)
library(ompr)
library(ROI)
library(ROI.plugin.glpk)
add_variable(MIPModel(), x, type = "continuous") \%>\%
 set_objective(x, sense = "max") \%>\%
 add_constraint(x <= 5) \%>\%
 solve_model(with_ROI(solver = "glpk", verbose = TRUE))

}
\references{
Kurt Hornik, David Meyer, Florian Schwendinger and Stefan Theussl (2016).
ROI: R Optimization Infrastructure. <https://CRAN.R-project.org/package=ROI>
}
